# components/agent-bff/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zuno Agent BFF API",
    "version": "1.0.0",
    "description": "Per-agent Backend-for-Frontend (ADR-0008) sitting between components/agent-frontend and the shared Agent Runtime. One BFF deployment per agent (v0: Tekos only) - AGENT_NAME selects which agent's Runtime path this instance calls. No public OpenShift Route: only components/agent-frontend calls this API, over its in-cluster Service. Authored as JSON rather than YAML so components/agent-bff's contract_test.go can parse it with encoding/json alone, preserving this component's deliberate zero-external-Go-dependency property (see README.md)."
  },
  "servers": [
    {
      "url": "http://tekos-bff.zuno-ai-run.svc.cluster.local:8080",
      "description": "In-cluster Service (Tekos deployment, v0's only agent)"
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz",
        "summary": "Liveness probe",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "text/plain": { "schema": { "type": "string", "example": "ok" } } }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "operationId": "chat",
        "summary": "Submit one chat turn for this agent",
        "description": "Validates the caller's bearer JWT and the agent_<AGENT_NAME> entitlement group (ADR-0040), then forwards the turn to the shared Agent Runtime (ADR-0032 identity propagation - the same validated token is forwarded, never a service credential). Returns a buffered JSON reply by default, or relays a Server-Sent Events stream byte-for-byte (ADR-0045) when the caller sends Accept: text/event-stream.",
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": false,
            "description": "text/event-stream requests the SSE variant of this response (ADR-0045) instead of the default synchronous JSON reply.",
            "schema": { "type": "string", "enum": ["application/json", "text/event-stream"] }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ChatRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Synchronous JSON reply (default, Content-Type: application/json), or a Server-Sent Events stream (Content-Type: text/event-stream) when the request's Accept header asked for one (ADR-0045). The SSE variant is a sequence of frames, in order: one 'start' frame (SseStartEvent), zero or more 'tool' frames (SseToolEvent), zero or more 'token' frames (SseTokenEvent), then exactly one terminal frame - either 'done' (SseDoneEvent) or 'error' (SseErrorEvent). Each frame's data payload matches the named component schema below. OpenAPI 3.0 has no first-class way to type a stream of discriminated event payloads under one response, so they are documented in this prose plus as ordinary schemas under components.schemas rather than via a non-standard vendor extension a validator may not recognize.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ChatResponse" } },
              "text/event-stream": { "schema": { "type": "string" } }
            }
          },
          "400": {
            "description": "Invalid request body (missing/empty message).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group (ADR-0040). The response never names which group was expected or which groups the caller actually has (Security considerations: \"document authorization failures without leaking policy internals\").",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "502": {
            "description": "The Agent Runtime was unreachable or returned an error.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/conversations": {
      "get": {
        "operationId": "listConversations",
        "summary": "List the caller's own conversations for this agent (ADR-0212)",
        "description": "Validates the caller's bearer JWT and the agent_<AGENT_NAME> entitlement group (ADR-0040), then forwards to the shared Agent Runtime. Starred conversations first, most recently updated first.",
        "parameters": [
          {
            "name": "starred",
            "in": "query",
            "required": false,
            "description": "true to return only starred conversations.",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "The caller's conversations.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Conversation" } } } }
          },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/conversations/{run_id}/transcript": {
      "get": {
        "operationId": "getTranscript",
        "summary": "Read a conversation's structured message history (ADR-0212)",
        "description": "Reused when reopening a conversation from the left-nav conversation list, to repopulate its exact prior message history.",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "The conversation's turns, oldest first.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TranscriptTurn" } } } }
          },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group, or this run_id belongs to a different subject.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "No workflow run found for this run_id.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/conversations/{run_id}": {
      "patch": {
        "operationId": "renameConversation",
        "summary": "Rename a conversation (ADR-0212)",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RenameRequest" } } }
        },
        "responses": {
          "200": { "description": "Renamed.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RenameResponse" } } } },
          "400": { "description": "Invalid request body (missing/empty title).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "No conversation found for this run_id owned by the caller (collapsed with 'not owned by the caller' - never confirms another subject's run_id exists).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "operationId": "archiveConversation",
        "summary": "Delete a conversation (ADR-0212 follow-up, soft-delete)",
        "description": "Hides the conversation from GET /api/conversations by setting archived_at - the underlying Agent Runtime checkpoint (the actual message history) is never deleted.",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Archived.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArchiveResponse" } } } },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "No conversation found for this run_id owned by the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/conversations/reorder": {
      "put": {
        "operationId": "reorderConversations",
        "summary": "Persist a drag-drop reorder of the caller's conversation list (ADR-0515)",
        "description": "The caller's full desired run_id order for this agent - each entry's index becomes its server-side sort_order. run_ids the caller doesn't own are silently skipped rather than failing the whole request.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReorderRequest" } } }
        },
        "responses": {
          "200": { "description": "Reordered.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReorderResponse" } } } },
          "400": { "description": "Invalid request body (empty run_ids).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/conversations/{run_id}/hard-delete": {
      "delete": {
        "operationId": "hardDeleteConversation",
        "summary": "Permanently delete a conversation (ADR-0515)",
        "description": "Irreversible - unlike DELETE /api/conversations/{run_id} (soft-delete/archive), this purges both the conversation's metadata row and its underlying LangGraph checkpoint/message history. The frontend requires an explicit user confirmation before calling this.",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HardDeleteResponse" } } } },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "No conversation found for this run_id owned by the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/conversations/{run_id}/star": {
      "put": {
        "operationId": "starConversation",
        "summary": "Star a conversation (ADR-0212)",
        "description": "The caller's own personal, private star - never shared with other users.",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Starred.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StarResponse" } } } },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "No conversation found for this run_id owned by the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "operationId": "unstarConversation",
        "summary": "Unstar a conversation (ADR-0212)",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Unstarred.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StarResponse" } } } },
          "401": { "description": "Missing, invalid or expired bearer token.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Token is valid but lacks the agent_<AGENT_NAME> entitlement group.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "No conversation found for this run_id owned by the caller.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "The Agent Runtime was unreachable or returned an error.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak-issued OIDC access token (RS256) - the same token the caller (components/agent-frontend) already received from the user's browser session and this service independently revalidates against Keycloak's JWKS (ADR-0013), never a service credential of the BFF's own. Never echoed back in any response body or schema below (Security considerations: \"never expose internal tokens in schemas\")."
      }
    },
    "schemas": {
      "ChatRequest": {
        "type": "object",
        "required": ["session_id", "message"],
        "properties": {
          "session_id": { "type": "string", "description": "Caller-generated correlation ID for this chat session." },
          "message": { "type": "string", "minLength": 1 },
          "project_id": { "type": "string", "description": "ADR-0209: scopes this turn to a project's durable, cross-session, cross-agent memory (knowledge.project). Optional - omitted or empty means no project memory is read or written for this turn. Propagated to the Agent Runtime exactly as received, following the same identity-propagation pattern as ADR-0032/0033 - this BFF never validates or resolves project membership itself." },
          "run_id": { "type": "string", "description": "ADR-0212: identifies the conversation to resume. Optional - omit to start a new conversation (the Agent Runtime mints a fresh run_id, returned in the SSE 'start' event); supply a prior run_id to resume it from its last checkpoint. Propagated to the Agent Runtime exactly as received, which enforces ownership - this BFF does not validate it itself." }
        }
      },
      "Citation": {
        "type": "object",
        "required": ["source", "title"],
        "properties": {
          "source": { "type": "string" },
          "title": { "type": "string" }
        }
      },
      "Image": {
        "type": "object",
        "description": "ADR-0415: a generate_image tool result for this turn (stable-diffusion-xl via OVHcloud AI Endpoints).",
        "required": ["data_base64", "mime_type", "alt"],
        "properties": {
          "data_base64": { "type": "string" },
          "mime_type": { "type": "string" },
          "alt": { "type": "string" }
        }
      },
      "ChatResponse": {
        "type": "object",
        "description": "ADR-0215: run_id/source_mode added to the synchronous JSON reply - previously only the SSE 'start' event exposed run_id, so a synchronous caller had no way to resume the conversation for a follow-up turn.",
        "required": ["reply", "citations", "images", "run_id", "source_mode"],
        "properties": {
          "reply": { "type": "string" },
          "citations": { "type": "array", "items": { "$ref": "#/components/schemas/Citation" } },
          "images": { "type": "array", "items": { "$ref": "#/components/schemas/Image" }, "description": "ADR-0415: empty unless this turn's agent/task is entitled to generate_image and the model chose to call it." },
          "run_id": { "type": "string", "description": "ADR-0212/ADR-0215: identifies this conversation - pass it as the request's run_id to resume with history on a later turn." },
          "source_mode": { "type": "string", "enum": ["indexed", "live", "both", "none"], "description": "ADR-0205: whether this reply used indexed knowledge, a live tool call, or both." }
        }
      },
      "Conversation": {
        "type": "object",
        "description": "ADR-0212: one entry of the caller's own conversation list.",
        "required": ["run_id", "title", "updated_at", "starred"],
        "properties": {
          "run_id": { "type": "string" },
          "title": { "type": "string" },
          "updated_at": { "type": "string", "format": "date-time" },
          "starred": { "type": "boolean" }
        }
      },
      "TranscriptTurn": {
        "type": "object",
        "description": "ADR-0212: one turn of a conversation's structured message history.",
        "required": ["role", "content", "ts"],
        "properties": {
          "role": { "type": "string", "enum": ["user", "assistant"] },
          "content": { "type": "string" },
          "ts": { "type": "string", "format": "date-time" },
          "images": { "type": "array", "items": { "$ref": "#/components/schemas/Image" }, "description": "ADR-0415: present only on an assistant turn that generated at least one image." }
        }
      },
      "RenameRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "minLength": 1 }
        }
      },
      "RenameResponse": {
        "type": "object",
        "required": ["run_id", "title"],
        "properties": {
          "run_id": { "type": "string" },
          "title": { "type": "string" }
        }
      },
      "StarResponse": {
        "type": "object",
        "required": ["starred"],
        "properties": {
          "starred": { "type": "boolean" }
        }
      },
      "ArchiveResponse": {
        "type": "object",
        "required": ["archived"],
        "properties": {
          "archived": { "type": "boolean" }
        }
      },
      "ReorderRequest": {
        "type": "object",
        "description": "ADR-0515: the caller's full desired conversation order for this agent.",
        "required": ["run_ids"],
        "properties": {
          "run_ids": { "type": "array", "items": { "type": "string" }, "minItems": 1 }
        }
      },
      "ReorderResponse": {
        "type": "object",
        "required": ["updated"],
        "properties": {
          "updated": { "type": "integer", "description": "Count of conversations actually reordered - run_ids the caller doesn't own are silently skipped." }
        }
      },
      "HardDeleteResponse": {
        "type": "object",
        "description": "ADR-0515: irreversible, unlike ArchiveResponse's soft-delete.",
        "required": ["deleted"],
        "properties": {
          "deleted": { "type": "boolean" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable reason - never a raw internal error or stack trace." }
        }
      },
      "SseStartEvent": {
        "type": "object",
        "description": "ADR-0045 request correlation - the first frame of every SSE stream.",
        "required": ["request_id"],
        "properties": { "request_id": { "type": "string", "format": "uuid" } }
      },
      "SseToolEvent": {
        "type": "object",
        "description": "Emitted around a live tool call (e.g. search_confluence) actually running.",
        "required": ["name", "status"],
        "properties": {
          "name": { "type": "string", "example": "search_confluence" },
          "status": { "type": "string", "enum": ["started", "finished"] }
        }
      },
      "SseTokenEvent": {
        "type": "object",
        "description": "One per model token/chunk.",
        "required": ["delta"],
        "properties": { "delta": { "type": "string" } }
      },
      "SseDoneEvent": {
        "type": "object",
        "description": "Terminal frame on success.",
        "required": ["citations"],
        "properties": { "citations": { "type": "array", "items": { "$ref": "#/components/schemas/Citation" } } }
      },
      "SseErrorEvent": {
        "type": "object",
        "description": "Terminal frame instead of done if the graph raised mid-stream.",
        "required": ["message"],
        "properties": { "message": { "type": "string" } }
      }
    }
  }
}